requirements.txt
aioredis==1.3.1
asgiref==3.3.4
async-timeout==3.0.1
asyncpg==0.23.0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
fastapi==0.65.2
fastapi-cache2==0.1.3.4
h11==0.12.0
hiredis==2.0.0
idna==2.10
pydantic==1.8.2
python-dateutil==2.8.1
requests==2.25.1
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.0
ujson==4.0.2
urllib3==1.26.5
uvicorn==0.14.0
